Permissions added

+++ src/main/java/edu/netcracker/messenger/user/controller/UserController.java
@@ -5,9 +5,12 @@ import edu.netcracker.messenger.user.exceptions.UserNotFoundException;
 import edu.netcracker.messenger.user.UserRepository;
 import edu.netcracker.messenger.user.views.UserPrivateView;
 import edu.netcracker.messenger.user.views.UserPublicView;
+import edu.netcracker.messenger.user.views.UserView;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -27,13 +30,19 @@ public class UserController {
      * @throws UserNotFoundException пользователей не найдено
      */
     @GetMapping("/all")
-    public @ResponseBody List<UserPublicView> allUsers() throws UserNotFoundException {
+    public @ResponseBody List<UserView> allUsers(Principal principal) throws UserNotFoundException {
         if (repository.findAll().isEmpty()) {
             throw new UserNotFoundException();
         }
-        List<UserPublicView> users = new ArrayList<>();
-        for (User user : repository.findAll()) {
-            users.add(new UserPublicView(user));
+        List<UserView> users = new ArrayList<>();
+        if (repository.findByUsername(principal.getName()).getAccountType().equals("ADMIN")) {
+            for (User user : repository.findAll()) {
+                users.add(new UserPrivateView(user));
+            }
+        } else {
+            for (User user : repository.findAll()) {
+                users.add(new UserPublicView(user));
+            }
         }
         return users;
     }
@@ -46,11 +55,15 @@ public class UserController {
      */
     @GetMapping("/{id}")
     public @ResponseBody
-    UserPrivateView getUserInfo(@PathVariable Long id) throws UserNotFoundException {
+    UserView getUserInfo(Principal principal, @PathVariable Long id) throws UserNotFoundException {
         if (repository.findById(id).isEmpty()) {
             throw new UserNotFoundException(id);
         }
-        return new UserPrivateView(repository.findById(id).get());
+        if (repository.findByUsername(principal.getName()).getId().equals(id) ||
+                repository.findByUsername(principal.getName()).getAccountType().equals("ADMIN"))  {
+            return new UserPrivateView(repository.findById(id).get());
+        }
+        return new UserPublicView(repository.findById(id).get());
     }
 
     /**
@@ -58,24 +71,20 @@ public class UserController {
      * @param id id пользователя
      * @return id удаленного пользователя в случае успеха
      * @throws UserNotFoundException пользователь с даннм id не найден
+     * @throws AccessDeniedException нет прав на удаление пользователя
      */
     @DeleteMapping("/{id}")
-    public @ResponseBody Long deleteUser(@PathVariable Long id) throws UserNotFoundException {
+    public @ResponseBody Long deleteUser(Principal principal, @PathVariable Long id)
+            throws UserNotFoundException, AccessDeniedException {
         if (repository.findById(id).isEmpty()) {
             throw new UserNotFoundException(id);
         }
+        if (!repository.findByUsername(principal.getName()).getId().equals(id) &&
+                !repository.findByUsername(principal.getName()).getAccountType().equals("ADMIN"))  {
+            throw new AccessDeniedException(String.format("You don't have permission to delete user with id: %d", id));
+        }
         repository.deleteById(id);
         return id;
-    }
-
-    /**
-     * Выполняет вход пользователя по id.
-     * @param id id пользователя
-     * @return подтверждение входа
-     */
-    @PostMapping("/login/{id}")
-    public @ResponseBody String loginUser(@PathVariable Long id) {
-        return String.format("Session started with user %d", id);
     }
 
     /**
